requirements.txt
gunicorn>=23.0.0
flasgger>=0.9.5
opencv-python>=4.9.0
pytesseract>=0.3.10
Pillow>=11.1.0
scikit-learn>=1.6.1
pandas>=2.2.3
numpy>=2.2.3
scipy>=1.15.2
joblib>=1.4.2
torch>=2.6.0
SQLAlchemy>=2.0.38
psycopg2>=2.9.10
pymongo>=4.6.0
tenacity>=9.0.0
rich>=13.9.4
loguru>=0.7.2
pytest-cov>=5.0.0
flake8>=7.1.2
black>=25.1.0
isort>=6.0.1
tox>=4.11.0
docker>=7.0.0
docker-compose>=1.29.2
pre-commit>=4.1.0
requests>=2.32.3
fastapi>=0.115.8
uvicorn>=0.27.1
python-dotenv>=1.0.1
